refactor(frontend): split playtest scene impl
